Consolidate packing / unpacking code
Mostly for safety and maintainability.  This makes it easy to see the overall
encoding scheme, between repr.rs and the representation tests in atom/mod.rs.

Thanks to the magic of LLVM's inlining and SROA, this code is faster on some
benchmarks and not significantly slower on any.

Also added debug_assert!s and changed the existing non-test assert!s.  I
believe they're all safe to omit in release builds.