Make tests pass on a 32-bit system.
Testable on 64-bit Linux with:

```
rustup target add i686-unknown-linux-gnu
cargo test --target i686-unknown-linux-gnu
```

(or similarly on anther 64-bit platforms),
assuming a linker and libc for this target are available on the system.

Leaving #162 open to add CI for this.
(Unfortunately Travis doesn’t use rustup out of the box.)